Implemented encryption

+++ Service/AccountService.cs
@@ -1,8 +1,10 @@
 ﻿using AuthSystem.Context;
+using AuthSystem.Security;
 using AuthSystem.Util.Constants;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.ListView;
 
 namespace AuthSystem
 {
@@ -17,7 +19,14 @@ namespace AuthSystem
 
         public List<Account> FindAll()
         {
-            return context.Accounts.ToList();
+            var accounts = context.Accounts.ToList();
+            var mappedAccounts = new List<Account>();
+            foreach (var account in accounts)
+            {
+                mappedAccounts.Add(MapAccount(account));
+            }
+
+            return mappedAccounts;
         }
 
         public Account FindById(int id)
@@ -29,7 +38,7 @@ namespace AuthSystem
                 throw new Exception(AppConstant.GetExceptionMessage("Account", "id", AppConstant.NOT_FOUND));
             }
 
-            return account;
+            return MapAccount(account);
         }
 
         public Account FindByUsername(string username)
@@ -41,48 +50,74 @@ namespace AuthSystem
                 throw new Exception(AppConstant.GetExceptionMessage("Account", "username", AppConstant.NOT_FOUND));
             }
 
-            return account;
+            return MapAccount(account);
         }
 
         public Account FindByEmail(string email)
         {
-            Account account = context.Accounts.FirstOrDefault(a => a.Email.Equals(email));
+            Account account = null;
+            foreach (var a in context.Accounts)
+            {
+                if (SecurityService.Decrypt(a.Email).Equals(email))
+                {
+                    account = a;
+                    break;
+                }
+            }
 
             if (account == null)
             {
                 throw new Exception(AppConstant.GetExceptionMessage("Account", "email", AppConstant.NOT_FOUND));
             }
 
-            return account;
+            return MapAccount(account);
         }
 
         public Account FindByPhoneNumber(string phoneNumber)
         {
-            Account account = context.Accounts.FirstOrDefault(a => a.PhoneNumber.Equals(phoneNumber));
+            Account account = null;
+            foreach (var a in context.Accounts)
+            {
+                if (SecurityService.Decrypt(a.PhoneNumber).Equals(phoneNumber))
+                {
+                    account = a;
+                    break;
+                }
+            }
 
             if (account == null)
             {
                 throw new Exception(AppConstant.GetExceptionMessage("Account", "phone number", AppConstant.NOT_FOUND));
             }
 
-            return account;
+            return MapAccount(account);
         }
 
         public void Add(Account account)
         {
+            account.Password = SecurityService.Encrypt(account.Password);
+            account.Email = SecurityService.Encrypt(account.Email);
+            account.PhoneNumber = SecurityService.Encrypt(account.PhoneNumber);
+
             context.Accounts.Add(account);
             context.SaveChanges();
         }
 
         public void Update(int id, Account account)
         {
-            var existingAccount = FindById(id);
+            var existingAccount = context.Accounts.Find(id);
+
+            if (existingAccount == null)
+            {
+                throw new Exception(AppConstant.GetExceptionMessage("Account", "id", AppConstant.NOT_FOUND));
+            }
+
             existingAccount.Username = account.Username;
-            existingAccount.Password = account.Password;
+            existingAccount.Password = SecurityService.Encrypt(account.Password);
             existingAccount.FirstName = account.FirstName;
             existingAccount.LastName = account.LastName;
-            existingAccount.Email = account.Email;
-            existingAccount.PhoneNumber = account.PhoneNumber;
+            existingAccount.Email = SecurityService.Encrypt(account.Email);
+            existingAccount.PhoneNumber = SecurityService.Encrypt(account.PhoneNumber);
             existingAccount.Role = account.Role != 0 ? account.Role : existingAccount.Role;
 
             context.SaveChanges();
@@ -90,8 +125,31 @@ namespace AuthSystem
 
         public void Remove(Account account)
         {
+            account = context.Accounts.Find(account.Id);
+
+            if (account == null)
+            {
+                throw new Exception(AppConstant.GetExceptionMessage("Account", "id", AppConstant.NOT_FOUND));
+            }
+
             context.Accounts.Remove(account);
             context.SaveChanges();
+        }
+
+        private Account MapAccount(Account account)
+        {
+            Account mappedAccount = new Account();
+            mappedAccount.Id = account.Id;
+            mappedAccount.Username = account.Username;
+            mappedAccount.Password = SecurityService.Decrypt(account.Password);
+            mappedAccount.FirstName = account.FirstName;
+            mappedAccount.LastName = account.LastName;
+            mappedAccount.Email = SecurityService.Decrypt(account.Email);
+            mappedAccount.PhoneNumber = SecurityService.Decrypt(account.PhoneNumber);
+            mappedAccount.Role = account.Role;
+            mappedAccount.Role1 = account.Role1;
+
+            return mappedAccount;
         }
     }
 }
